deletes a ton of extraneous code

+++ src/app/file/views.py
@@ -17,17 +17,13 @@ def verify_token(token):
     resp = requests.get(
         "http://localhost:7000/auth", headers={"Authorization": f"Bearer {token}"}
     )
-    if resp.status_code == HTTPStatus.OK:
-        return token
-    else:
-        return None
+    return token if resp.status_code == HTTPStatus.OK else None
 
 
-@file_bp.route("/file", methods=["GET"])
+@file_bp.route("/file<path:string>", methods=["GET"])
 @token_auth.login_required
-def read():
-    file_path = request.args.get("path")
-    if None in (file_path):
+def read(file_path):
+    if file_path == None:
         return Response(
             "missing path from query params",
             status=HTTPStatus.BAD_REQUEST,
@@ -42,9 +38,15 @@ def read():
 
     matches = [f for f in os.listdirs(abs_dir) if f.startswith(file_name)]
 
-    if len(matches) == 0 or len(matches) < 2:
+    if len(matches) == 0:
         return Response(
-            f"There were too many or two few matches",
+            f"There were no matches",
+            status=HTTPStatus.INTERNAL_SERVER_ERROR,
+        )
+
+    if len(matches) > 2:
+        return Response(
+            f"There were too many matches",
             status=HTTPStatus.INTERNAL_SERVER_ERROR,
         )
 
@@ -55,7 +57,7 @@ def read():
     return jsonify(filePath=file_path, fileContent=content)
 
 
-@file_bp.route("/update", methods=["PUT"])
+@file_bp.route("/file", methods=["PUT"])
 @token_auth.login_required
 def update():
     body = request.json
@@ -71,7 +73,34 @@ def update():
             f"file path was missing. {file_path}", status=HTTPStatus.BAD_REQUEST
         )
 
+    if not os.path.exists(file_path):
+        return Response("The file to update does not exist. Please use /file POST to create.", status=HTTPStatus.INTERNAL_SERVER_ERROR)
+
     with open(file_path, "w", newline="\n") as my_file:
+        my_file.write(content)
+
+    return jsonify(filePath=file_path, fileContent=content)
+
+@file_bp.route("/file", methods=["POST"])
+@token_auth.login_required
+def create():
+    body = request.json
+    file_path, content = body.get("filePath"), body.get("fileContent")
+
+    if content in [None, ""]:
+        return Response(
+            f"content was missing. {content}", status=HTTPStatus.BAD_REQUEST
+        )
+
+    if file_path in [None, ""]:
+        return Response(
+            f"file path was missing. {file_path}", status=HTTPStatus.BAD_REQUEST
+        )
+
+    if os.path.exists(file_path):
+        return Response("The file to create already exists. Please use /file PUT to update.", status=HTTPStatus.INTERNAL_SERVER_ERROR)
+
+    with open(file_path, "x", newline="\n") as my_file:
         my_file.write(content)
 
     return jsonify(filePath=file_path, fileContent=content)
